mv GenericSeries, SeriesParam
