+++ weldx/core/__init__.py
@@ -1,2 +1,3 @@
+from .generic_series import GenericSeries
 from .math_expression import MathematicalExpression
 from .spatial_series import SpatialSeries
